rename wallet messages
