+++ packages/wallet/src/redux/sagas/__tests__/message-listener.test.ts
@@ -1,5 +1,5 @@
 import { put } from 'redux-saga/effects';
-import * as incoming from 'wallet-client/lib/wallet-messages';
+import * as incoming from 'wallet-client/lib/messages-to-wallet';
 
 import { messageListener } from '../message-listener';
 import * as actions from '../../actions';
